// src/app/bff/enums/prod/prodIdEnum.ts
/**
 * @description 商品代碼 (enum)
 */
export enum ProdIdEnum {
  /** 汽車險 */
  Auto = 'CAR001',
  /** 機車險 */
  Moto = 'MOTOR001',
  /** 住火險 */
  Fire = 'FIRE001',
  /** (投保證明) 住火險 */
  Fire011 = 'FIRE011',
  /** (投保證明) 住綜險 */
  Fire012 = 'FIRE012',
  /** 寵物險 */
  Pet = 'PET001',
  /** 個人綜合險(一階、二階皆無此險種，僅舊網投有) */
  Psn = 'PSN001',
  /** 個人傷害險 */
  Psninj = 'PSNINJ001',
  /** 旅遊險 */
  Travel = 'TRAVEL001',
  /** 旅行平安保險 */
  Travel011 = 'TRAVEL011',
  /** 登山綜合保險 */
  Travel012 = 'TRAVEL012',
  /** 郵輪旅遊綜合保險 */
  Travel013 = 'TRAVEL013',
  /** 海域活動綜合保險 */
  Travel014 = 'TRAVEL014',
  /** 旅遊泡泡綜合保險 */
  Travel015 = 'TRAVEL015',
}